codegen: Extract codegen context into a submodule
