+++ rustler_codegen/src/lib.rs
@@ -10,7 +10,7 @@ extern crate syn;
 #[macro_use]
 extern crate quote;
 
-use syn::{Lit, Meta, NestedMeta};
+mod context;
 
 mod ex_struct;
 mod map;
@@ -25,136 +25,6 @@ enum RustlerAttr {
     Decode,
     Module(String),
     Tag(String),
-}
-
-#[derive(Debug)]
-struct Context<'a> {
-    attrs: Vec<RustlerAttr>,
-    ident: &'a proc_macro2::Ident,
-    ident_with_lifetime: proc_macro2::TokenStream,
-}
-
-impl<'a> Context<'a> {
-    fn from_ast(ast: &'a syn::DeriveInput) -> Self {
-        let mut attrs: Vec<_> = ast
-            .attrs
-            .iter()
-            .map(Context::get_rustler_attrs)
-            .flatten()
-            .collect();
-
-        //
-        // Default: generate encoder and decoder
-        //
-        if !Context::encode_decode_attr_set(&attrs) {
-            attrs.push(RustlerAttr::Encode);
-            attrs.push(RustlerAttr::Decode);
-        }
-
-        let has_lifetime = match ast.generics.lifetimes().count() {
-            0 => false,
-            1 => true,
-            _ => panic!("Struct can only have one lifetime argument"),
-        };
-
-        let ident = &ast.ident;
-        let ident_with_lifetime = if has_lifetime {
-            quote! { #ident <'a> }
-        } else {
-            quote! { #ident }
-        };
-
-        Self {
-            attrs,
-            ident,
-            ident_with_lifetime,
-        }
-    }
-
-    fn encode(&self) -> bool {
-        self.attrs.iter().any(|attr| match attr {
-            RustlerAttr::Encode => true,
-            _ => false,
-        })
-    }
-
-    fn decode(&self) -> bool {
-        self.attrs.iter().any(|attr| match attr {
-            RustlerAttr::Decode => true,
-            _ => false,
-        })
-    }
-
-    fn encode_decode_attr_set(attrs: &[RustlerAttr]) -> bool {
-        attrs.iter().any(|attr| match attr {
-            RustlerAttr::Encode => true,
-            RustlerAttr::Decode => true,
-            _ => false,
-        })
-    }
-
-    fn get_rustler_attrs(attr: &syn::Attribute) -> Vec<RustlerAttr> {
-        attr.path
-            .segments
-            .iter()
-            .filter_map(|segment| {
-                let meta = attr.parse_meta().expect("can parse meta");
-                match segment.ident.to_string().as_ref() {
-                    "rustler" => Context::parse_rustler(&meta),
-                    "tag" => Context::try_parse_tag(&meta),
-                    "module" => Context::try_parse_module(&meta),
-                    _ => None,
-                }
-            })
-            .flatten()
-            .collect()
-    }
-
-    fn parse_rustler(meta: &Meta) -> Option<Vec<RustlerAttr>> {
-        if let Meta::List(ref list) = meta {
-            return Some(
-                list.nested
-                    .iter()
-                    .map(Context::parse_nested_rustler)
-                    .collect(),
-            );
-        }
-
-        panic!("Expected encode and/or decode in rustler attribute");
-    }
-
-    fn parse_nested_rustler(nested: &NestedMeta) -> RustlerAttr {
-        if let NestedMeta::Meta(ref meta) = nested {
-            if let Meta::Path(ref path) = meta {
-                match path.segments[0].ident.to_string().as_ref() {
-                    "encode" => return RustlerAttr::Encode,
-                    "decode" => return RustlerAttr::Decode,
-                    other => panic!("Unexpected literal {}", other),
-                }
-            }
-        }
-
-        panic!("Expected encode and/or decode in rustler attribute");
-    }
-
-    fn try_parse_tag(meta: &Meta) -> Option<Vec<RustlerAttr>> {
-        if let Meta::NameValue(ref name_value) = meta {
-            if let Lit::Str(ref tag) = name_value.lit {
-                return Some(vec![RustlerAttr::Tag(tag.value())]);
-            }
-        }
-        panic!("Cannot parse module")
-    }
-
-    fn try_parse_module(meta: &Meta) -> Option<Vec<RustlerAttr>> {
-        if let Meta::NameValue(name_value) = meta {
-            if let Lit::Str(ref module) = name_value.lit {
-                let ident = format!("Elixir.{}", module.value());
-                return Some(vec![RustlerAttr::Module(ident)]);
-            }
-        }
-        panic!("Cannot parse tag")
-    }
 }
 
 /// Implementation of the `NifStruct` macro that lets the user annotate a struct that will
