+++ rustler_codegen/src/map.rs
@@ -2,7 +2,7 @@ use proc_macro2::{Span, TokenStream};
 
 use syn::{self, Data, Field, Ident};
 
-use super::Context;
+use super::context::Context;
 
 pub fn transcoder_decorator(ast: &syn::DeriveInput) -> TokenStream {
     let ctx = Context::from_ast(ast);
