+++ rustler_codegen/src/untagged_enum.rs
@@ -2,7 +2,7 @@ use proc_macro2::TokenStream;
 
 use syn::{self, Data, Fields, Variant};
 
-use super::Context;
+use super::context::Context;
 
 pub fn transcoder_decorator(ast: &syn::DeriveInput) -> TokenStream {
     let ctx = Context::from_ast(ast);
